requirements.txt
streamlit
fastapi
uvicorn
langchain
langchain-groq
langchain-community
langgraph
tavily-python
python-dotenv
pydantic
requests